Fixed TarOutputStream for files bigger than 2GB
git-svn-id: https://svn.apache.org/repos/asf/commons/sandbox/compress/trunk@670243 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/compress/archivers/tar/TarOutputStream.java
@@ -62,8 +62,8 @@ public final class TarOutputStream extends FilterOutputStream
     private byte[] m_assemBuf;
     private int m_assemLen;
     private TarBuffer m_buffer;
-    private int m_currBytes;
-    private int m_currSize;
+    private long m_currBytes;
+    private long m_currSize;
 
     private byte[] m_oneBuf;
     private byte[] m_recordBuf;
@@ -262,7 +262,7 @@ public final class TarOutputStream extends FilterOutputStream
         }
         else
         {
-            m_currSize = (int)entry.getSize();
+            m_currSize = entry.getSize();
         }
     }
